+++ src/export/copying.rs
@@ -5,11 +5,13 @@ use colored::Colorize;
 use derive_new::new;
 use crate::export::structure::OutputStrategy;
 use crate::model::asset::ExportAsset;
+use crate::model::uti::Uti;
 
 
 #[derive(new)]
 pub struct CopyOperation {
     pub source_path: PathBuf,
+    pub file_type: &'static Uti,
     pub output_filename: String,
     pub output_filename_suffix: Option<String>,
     pub output_folder: Option<PathBuf>,
@@ -21,9 +23,10 @@ impl CopyOperation {
             .join(self.output_folder.clone().unwrap_or(PathBuf::new()))
             .join(
                 format!(
-                    "{}{}",
+                    "{}{}.{}",
                     self.output_filename,
-                    self.output_filename_suffix.clone().unwrap_or("".to_string())
+                    self.output_filename_suffix.clone().unwrap_or("".to_string()),
+                    self.file_type.extension
                 )
             )
     }
@@ -31,19 +34,63 @@ impl CopyOperation {
 
 
 pub trait CopyOperationFactory {
-    fn build(&self, asset: &ExportAsset) -> Vec<CopyOperation>;
+    fn build(&self, asset: &ExportAsset) -> Result<Vec<CopyOperation>, String>;
 }
 
 #[derive(new)]
 pub struct OriginalsCopyOperationFactory;
 impl CopyOperationFactory for OriginalsCopyOperationFactory {
-    fn build(&self, asset: &ExportAsset) -> Vec<CopyOperation> {
-        vec![CopyOperation::new(
-            asset.path(),
-            asset.filename.clone(),
+    fn build(&self, asset: &ExportAsset) -> Result<Vec<CopyOperation>, String> {
+        let operation = CopyOperation::new(
+            asset.get_path(),
+            asset.original_uti,
+            asset.uuid.clone(),
             None,
             None,
-        )]
+        );
+        Ok(vec![operation])
+    }
+}
+
+#[derive(new)]
+pub struct DerivatesCopyOperationFactory;
+impl CopyOperationFactory for DerivatesCopyOperationFactory {
+    fn build(&self, asset: &ExportAsset) -> Result<Vec<CopyOperation>, String> {
+        let operations = if asset.has_adjustments {
+            vec![
+                CopyOperation::new(
+                    asset.get_derivate_path().ok_or("No derivate path")?,
+                    asset.derivate_uti,
+                    asset.uuid.clone(),
+                    Some("_edited".to_string()),
+                    None,
+                )
+            ]
+        } else {
+            vec![]
+        };
+        Ok(operations)
+    }
+}
+
+#[derive(new)]
+pub struct CombiningCopyOperationFactory {
+    factories: Vec<Box<dyn CopyOperationFactory>>,
+}
+
+impl CopyOperationFactory for CombiningCopyOperationFactory {
+    fn build(&self, asset: &ExportAsset) -> Result<Vec<CopyOperation>, String> {
+        let mut operations = self.factories
+            .iter()
+            .map(|factory| factory.build(asset))
+            .collect::<Result<Vec<Vec<_>>, _>>()?
+            .into_iter()
+            .flatten()
+            .collect::<Vec<CopyOperation>>();
+
+        operations.sort_by_key(|op| op.source_path.to_string_lossy().into_owned());
+
+        Ok(operations)
     }
 }
 
@@ -52,19 +99,23 @@ pub struct FilenameRestoringCopyOperationFactoryDecorator {
     inner: Box<dyn CopyOperationFactory>,
 }
 impl CopyOperationFactory for FilenameRestoringCopyOperationFactoryDecorator {
-    fn build(&self, asset: &ExportAsset) -> Vec<CopyOperation> {
+    fn build(&self, asset: &ExportAsset) -> Result<Vec<CopyOperation>, String> {
         self.inner
-            .build(asset)
+            .build(asset)?
             .into_iter()
             .map(|op| {
-                CopyOperation::new(
-                    op.source_path,
-                    asset.original_filename.clone(),
-                    op.output_filename_suffix,
-                    op.output_folder,
-                )
+                let original_filename_stem = PathBuf::from(&asset.original_filename)
+                    .file_stem()
+                    .ok_or("Failed to get file stem")?
+                    .to_string_lossy()
+                    .to_string();
+
+                Ok(CopyOperation {
+                    output_filename: original_filename_stem,
+                    ..op
+                })
             })
-            .collect()
+            .collect::<Result<Vec<CopyOperation>, String>>()
     }
 }
 
@@ -74,19 +125,41 @@ pub struct OutputStructureCopyOperationFactoryDecorator {
     strategy: Box<dyn OutputStrategy>,
 }
 impl CopyOperationFactory for OutputStructureCopyOperationFactoryDecorator {
-    fn build(&self, asset: &ExportAsset) -> Vec<CopyOperation> {
-        self.inner
-            .build(asset)
+    fn build(&self, asset: &ExportAsset) -> Result<Vec<CopyOperation>, String> {
+        let operations = self.inner
+            .build(asset)?
             .into_iter()
             .map(|op| {
-                CopyOperation::new(
-                    op.source_path,
-                    op.output_filename,
-                    op.output_filename_suffix,
-                    self.strategy.get_relative_output_dir(asset).ok(),
-                )
+                CopyOperation {
+                    output_folder: self.strategy.get_relative_output_dir(asset).ok(),
+                    ..op
+                }
             })
-            .collect()
+            .collect();
+
+        Ok(operations)
+    }
+}
+
+#[derive(new)]
+pub struct SuffixSettingCopyOperationFactoryDecorator {
+    inner: Box<dyn CopyOperationFactory>,
+    suffix: String,
+}
+impl CopyOperationFactory for SuffixSettingCopyOperationFactoryDecorator {
+    fn build(&self, asset: &ExportAsset) -> Result<Vec<CopyOperation>, String> {
+        let operations = self.inner
+            .build(asset)?
+            .into_iter()
+            .map(|op| {
+                CopyOperation {
+                    output_filename_suffix: Some(self.suffix.clone()),
+                    ..op
+                }
+            })
+            .collect();
+
+        Ok(operations)
     }
 }